Start converting the skaters table
